+++ api-parent/spi/src/main/java/com/atlassian/plugin/connect/spi/product/ProductAccessor.java
@@ -1,5 +1,7 @@
 package com.atlassian.plugin.connect.spi.product;
 
+import com.atlassian.extras.api.ProductLicense;
+import com.atlassian.fugue.Option;
 import com.atlassian.plugin.connect.spi.host.HostProperties;
 import com.atlassian.plugin.web.Condition;
 
@@ -27,4 +29,6 @@ public interface ProductAccessor extends HostProperties
     Map<String, Class<? extends Condition>> getConditions();
 
     boolean needsAdminPageNameEscaping();
+
+    Option<ProductLicense> getProductLicense();
 }