+++ plugin/src/main/java/com/atlassian/plugin/connect/plugin/capabilities/provider/DefaultWebItemModuleProvider.java
@@ -93,6 +93,7 @@ public class DefaultWebItemModuleProvider implements WebItemModuleProvider
 
             iFrameRenderStrategyRegistry.register(connectAddonBean.getKey(), bean.getKey(connectAddonBean), iFrameRenderStrategy);
             iFrameRenderStrategyRegistry.register(connectAddonBean.getKey(), bean.getRawKey(), RAW_CLASSIFIER, iFrameRenderStrategy);
+            iFrameRenderStrategyRegistry.register(connectAddonBean.getKey(), bean.getRawKey(), iFrameRenderStrategy);
         }
 
         return descriptors;