+++ plugin/src/main/resources/js/confluence/macro/editor-rpc.js
@@ -1,21 +1,25 @@
-_AP.define("confluence/macro/editor/rpc", ["_dollar", "confluence/macro/editor", "_rpc"], function($, editor, rpc) {
-    rpc.extend(function () {
-        return {
-            internals: {
-                saveMacro: function (updatedParams, updatedMacroBody) {
-                    editor.saveMacro(updatedParams, updatedMacroBody);
-                },
-                closeMacroEditor: function () {
-                    editor.close();
-                },
-                getMacroData: function (callback) {
-                    editor.getMacroData(callback);
-                },
-                getMacroBody: function (callback) {
-                    editor.getMacroBody(callback);
+(function($, require){
+    "use strict";
+
+    require(["ac/confluence/macro/editor", 'connect-host'], function(editor, _AP) {
+        _AP.extend(function () {
+            return {
+                internals: {
+                    saveMacro: function (updatedParams, updatedMacroBody) {
+                        editor.saveMacro(updatedParams, updatedMacroBody);
+                    },
+                    closeMacroEditor: function () {
+                        editor.close();
+                    },
+                    getMacroData: function (callback) {
+                        editor.getMacroData(callback);
+                    },
+                    getMacroBody: function (callback) {
+                        editor.getMacroBody(callback);
+                    }
                 }
-            }
-        };
+            };
+        });
     });
 
-});
+}(AJS.$, require));