+++ plugin/src/main/resources/js/jira/events.js
@@ -1,24 +1,26 @@
-_AP.define("jira/events", ["_dollar", "_rpc"], function($, rpc) {
-
-    var module = {
-        refreshIssuePage: function(){
-            try {
-                JIRA.trigger(JIRA.Events.REFRESH_ISSUE_PAGE, [JIRA.Issue.getIssueId()]);
-            } catch (e){
-                throw new Error("Failed to refresh the issue page");
-            }
-        }
-    };
-
-    rpc.extend(function () {
-        return {
-            internals: {
-                triggerJiraEvent: function () {
-                    module.refreshIssuePage();
+(function(define, $){
+    "use strict";
+    define("ac/jira/events", ['connect-host'], function(_AP){
+        var module = {
+            refreshIssuePage: function(){
+                try {
+                    JIRA.trigger(JIRA.Events.REFRESH_ISSUE_PAGE, [JIRA.Issue.getIssueId()]);
+                } catch (e){
+                    throw new Error("Failed to refresh the issue page");
                 }
             }
         };
+
+        _AP.extend(function () {
+            return {
+                internals: {
+                    triggerJiraEvent: function () {
+                        module.refreshIssuePage();
+                    }
+                }
+            };
+        });
+        return module;
     });
 
-    return module;
-});
+})(define, AJS.$);