+++ plugin/src/test/java/com/atlassian/plugin/connect/test/plugin/capabilities/module/ContentMacroTest.java
@@ -4,10 +4,10 @@ import com.atlassian.confluence.content.render.xhtml.ConversionContext;
 import com.atlassian.confluence.core.ContentEntityObject;
 import com.atlassian.confluence.macro.MacroExecutionException;
 import com.atlassian.plugin.connect.modules.beans.DynamicContentMacroModuleBean;
+import com.atlassian.plugin.connect.modules.beans.nested.MacroRenderModesBean;
 import com.atlassian.plugin.connect.plugin.capabilities.module.macro.DynamicContentMacro;
-import com.atlassian.plugin.connect.plugin.capabilities.module.macro.MacroModuleContextExtractor;
+import com.atlassian.plugin.connect.plugin.capabilities.module.macro.RemoteMacroRenderer;
 import com.atlassian.plugin.connect.plugin.capabilities.util.MacroEnumMapper;
-import com.atlassian.plugin.connect.plugin.iframe.render.strategy.IFrameRenderStrategy;
 import com.atlassian.plugin.connect.spi.module.IFrameContext;
 import com.atlassian.plugin.connect.test.plugin.capabilities.testobjects.ContentEntityForTests;
 import com.atlassian.sal.api.user.UserKey;
@@ -21,6 +21,7 @@ import org.mockito.ArgumentMatcher;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
+import java.util.HashMap;
 import java.util.Map;
 
 import static com.atlassian.plugin.connect.modules.beans.DynamicContentMacroModuleBean.newDynamicContentMacroModuleBean;
@@ -32,8 +33,7 @@ public class ContentMacroTest
 {
     @Mock private UserManager userManager;
     @Mock private ConversionContext conversionContext;
-    @Mock private IFrameRenderStrategy iFrameRenderStrategy;
-    @Mock private MacroModuleContextExtractor macroModuleContextExtractor;
+    @Mock private RemoteMacroRenderer remoteMacroRenderer;
 
     @Before
     public void beforeEachTest()
@@ -74,8 +74,12 @@ public class ContentMacroTest
                 .withHeight(height)
                 .build();
 
-        DynamicContentMacro macro = new DynamicContentMacro(MacroEnumMapper.map(bean.getBodyType()),
-                MacroEnumMapper.map(bean.getOutputType()), iFrameRenderStrategy, macroModuleContextExtractor);
+        DynamicContentMacro macro = new DynamicContentMacro(
+//                addOnKey, moduleKey,
+                "addon-key", "module-key",
+                MacroEnumMapper.map(bean.getBodyType()),
+                MacroEnumMapper.map(bean.getOutputType()), remoteMacroRenderer,
+                MacroRenderModesBean.newMacroRenderModesBean().build());
         macro.execute(Maps.<String, String>newHashMap(), "some macro content", conversionContext);
     }