+++ docs/public/resources/release-notes.md
@@ -32,6 +32,13 @@ or Confluence Cloud with Atlassian Connect as follows:
 <div id="runproduct-vagrant" class="aui-expander-content">
     <span data-include="../assets/includes/runproduct-vagrant.html">Loading...</span>
 </div>
+## [1.1.20](../release-notes/1-1-0.html#1.1.20)
+* JIRA: Whitelisted endpoints for thumbnails and user avatars [AC-1472](https://ecosystem.atlassian.net/browse/AC-1472)
+* JIRA: Added support for aliases in Entity Properties [ACJIRA-250](https://ecosystem.atlassian.net/browse/ACJIRA-250)
+* Improved REST API for license information [AC-1370](https://ecosystem.atlassian.net/browse/AC-1370)
+
+## [1.1.18](../release-notes/1-1-0.html#1.1.18)
+* Fixed: Confluence macro editor replaces page content after it gets closed in Internet Explorer
 
 ## [1.1.17](../release-notes/1-1-0.html#1.1.17)
 * Confluence: fixed a bug that made chromeless dialogs fail to display [AC-1449](https://ecosystem.atlassian.net/browse/AC-1449)